test(org): per-group npm scripts + split the largest domain test file
(B) Grouped test scripts: package.json gains test:domain / test:application
/ test:infrastructure / test:interface / test:passages / test:e2e, each
building then running only that subtree (tests/run.mjs already accepted a
root arg). Iterating on one area no longer requires the full ~1.8k suite;
default `npm test` is unchanged.

(A) Split tests/domain/Request.test.ts (1158 lines, 69 tests) into focused
files sharing one fixture module:
  - _requestHelpers.ts  — shared fixed clock `d` + `mkReq` builder
  - RequestId.test.ts   — id format/validation
  - RequestSlices294.test.ts — the self-contained #294 per-executor
    slice-closure block (helper + ~25 tests)
  - Request.test.ts     — trimmed core lifecycle / serialization / stamps
Same 69 tests, no behavior change. Exemplar for the remaining large files;
ax.test.ts is deliberately left untouched to avoid colliding with #420.

Full suite 1788 pass / 0 fail.
