+++ package.json
@@ -27,6 +27,12 @@
     "prepare": "tsc",
     "prepack": "npm run build",
     "test": "tsc && node tests/run.mjs",
+    "test:domain": "tsc && node tests/run.mjs dist/tests/domain",
+    "test:application": "tsc && node tests/run.mjs dist/tests/application",
+    "test:infrastructure": "tsc && node tests/run.mjs dist/tests/infrastructure",
+    "test:interface": "tsc && node tests/run.mjs dist/tests/interface",
+    "test:passages": "tsc && node tests/run.mjs dist/tests/passages",
+    "test:e2e": "tsc && node tests/run.mjs dist/tests/e2e",
     "typecheck": "tsc --noEmit",
     "gate": "node bin/gate.mjs",
     "guild": "node bin/guild.mjs",
